fix color issues
