+++ ENIGMAsystem/SHELL/Graphics_Systems/OpenGL3/GL3ModelStruct.h
@@ -21,7 +21,7 @@
 #include "Graphics_Systems/General/GSprimitives.h"
 #include "Graphics_Systems/General/GSmatrix.h"
 #include "Graphics_Systems/General/GSmath.h"
-#include "Graphics_Systems/General/GSvertex.h"
+#include "Graphics_Systems/General/GSvertex_impl.h"
 #include "Graphics_Systems/General/GScolor_macros.h"
 #include "Bridges/General/GL3Context.h" //Needed to get if bound texture == -1
 #include "GLSLshader.h"
@@ -139,25 +139,6 @@ namespace enigma {
   		}
   	}
   }
-
-  template<int x> struct intmatch { };
-  template<int x> struct uintmatch { };
-  template<> struct intmatch<1>   { typedef int8_t type;  };
-  template<> struct intmatch<2>   { typedef int16_t type; };
-  template<> struct intmatch<4>   { typedef int32_t type; };
-  template<> struct intmatch<8>   { typedef int64_t type; };
-  template<> struct uintmatch<1>  { typedef uint8_t type;  };
-  template<> struct uintmatch<2>  { typedef uint16_t type; };
-  template<> struct uintmatch<4>  { typedef uint32_t type; };
-  template<> struct uintmatch<8>  { typedef uint64_t type; };
-  typedef uintmatch<sizeof(gs_scalar)>::type color_t;
-  union VertexElement {
-  	color_t d;
-  	gs_scalar f;
-
-  	VertexElement(gs_scalar v): f(v) {}
-  	VertexElement(color_t v): d(v) {}
-  };
 
   //NOTE: This class handles batching, indexing, and other optimization for you and is very very efficient.
   class Mesh
